Move to ClassLib and remove semanticversion.
