--- Source/ChocolateyGui.Shared/ChocolateyGui.Shared.projitems
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>eb4b7c3f-4fdc-474b-906f-d2b274947c11</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>ChocolateyGui.Shared</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)IIpcChocolateyService.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IIpcServiceCallbacks.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\ChocolateyFeature.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\ChocolateySetting.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\ChocolateySource.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\Package.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\PackageOperationResult.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\PackageSearchOptions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\PackageResults.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\SemanticVersion.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\SemanticVersionTypeConverter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\StreamingLogLevel.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Models\StreamingLogMessage.cs" />
-  </ItemGroup>
-</Project>
