+++ Source/ChocolateyGui.Shared/Models/ChocolateySetting.cs
@@ -15,11 +15,9 @@ namespace ChocolateyGui.Models
         public string Key { get; set; }
 
         [DataMember]
-
         public string Value { get; set; }
 
         [DataMember]
-
         public string Description { get; set; }
     }
 }
