+++ Source/ChocolateyGui.Shared/Models/ChocolateySource.cs
