+++ Source/ChocolateyGui.Shared/Models/Package.cs
@@ -6,6 +6,7 @@
 
 using System;
 using System.Runtime.Serialization;
+using NuGet;
 
 namespace ChocolateyGui.Models
 {
@@ -16,7 +17,6 @@ namespace ChocolateyGui.Models
         public string[] Authors { get; set; }
 
         [DataMember]
-
         public string Copyright { get; set; }
 
         [DataMember]
@@ -24,115 +24,87 @@ namespace ChocolateyGui.Models
         public string Dependencies { get; set; }
 
         [DataMember]
-
         public string Description { get; set; }
 
         [DataMember]
-
         public int DownloadCount { get; set; }
 
         [DataMember]
-
         public string GalleryDetailsUrl { get; set; }
 
         [DataMember]
-
         public string IconUrl { get; set; }
 
         [DataMember]
-
         public string Id { get; set; }
 
         [DataMember]
-
         public bool IsAbsoluteLatestVersion { get; set; }
 
         [DataMember]
-
         public bool IsInstalled { get; set; }
 
         [DataMember]
-
         public bool IsPinned { get; set; }
 
         [DataMember]
-
         public bool IsLatestVersion { get; set; }
 
         [DataMember]
-
         public bool IsPrerelease { get; set; }
 
         [DataMember]
-
         public string Language { get; set; }
 
         [DataMember]
-
         public string LatestVersion { get; set; }
 
         [DataMember]
-
         public string LicenseUrl { get; set; }
 
         [DataMember]
-
         public string[] Owners { get; set; }
 
         [DataMember]
-
         public string PackageHash { get; set; }
 
         [DataMember]
-
         public string PackageHashAlgorithm { get; set; }
 
         [DataMember]
-
         public long PackageSize { get; set; }
 
         [DataMember]
-
         public string ProjectUrl { get; set; }
 
         [DataMember]
-
         public DateTimeOffset Published { get; set; }
 
         [DataMember]
-
         public string ReleaseNotes { get; set; }
 
         [DataMember]
-
         public string ReportAbuseUrl { get; set; }
 
         [DataMember]
-
         public string RequireLicenseAcceptance { get; set; }
 
         [DataMember]
-
         public Uri Source { get; set; }
 
         [DataMember]
-
         public string Summary { get; set; }
 
         [DataMember]
-
         public string Tags { get; set; }
 
         [DataMember]
-
         public string Title { get; set; }
 
         [DataMember]
-
-        public string Version { get; set; }
+        public SemanticVersion Version { get; set; }
 
         [DataMember]
-
         public int VersionDownloadCount { get; set; }
     }
 }
