+++ Source/ChocolateyGui.Shared/Models/PackageSearchOptions.cs
