+++ Source/ChocolateyGui.Shared/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="chocolatey.lib" version="0.10.5" targetFramework="net452" />
+  <package id="log4net" version="2.0.3" targetFramework="net452" />
+</packages>
