+++ Source/ChocolateyGui/Models/Messages/PackageChangedMessage.cs
@@ -4,6 +4,8 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
+using NuGet;
+
 namespace ChocolateyGui.Models.Messages
 {
     public enum PackageChangeType
