+++ Source/ChocolateyGui/Models/Messages/PackageHasUpdateMessage.cs
@@ -4,6 +4,8 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
+using NuGet;
+
 namespace ChocolateyGui.Models.Messages
 {
     public class PackageHasUpdateMessage
