+++ Source/ChocolateyGui/ViewModels/Items/IPackageViewModel.cs
@@ -6,6 +6,7 @@
 
 using System;
 using System.Threading.Tasks;
+using NuGet;
 
 namespace ChocolateyGui.ViewModels.Items
 {
