+++ Source/ChocolateyGui/ViewModels/Items/PackageViewModel.cs
@@ -13,6 +13,7 @@ using ChocolateyGui.Base;
 using ChocolateyGui.Models.Messages;
 using ChocolateyGui.Properties;
 using ChocolateyGui.Services;
+using NuGet;
 using Action = System.Action;
 using MemoryCache = System.Runtime.Caching.MemoryCache;
