+++ Source/ChocolateyGui/packages.config
@@ -10,7 +10,9 @@
   <package id="CefSharp.Common" version="57.0.0" targetFramework="net452" />
   <package id="CefSharp.OffScreen" version="57.0.0" targetFramework="net452" />
   <package id="CefSharp.Wpf" version="57.0.0" targetFramework="net452" />
+  <package id="chocolatey.lib" version="0.10.5" targetFramework="net452" />
   <package id="LiteDB" version="3.1.0" targetFramework="net452" />
+  <package id="log4net" version="2.0.3" targetFramework="net452" />
   <package id="MahApps.Metro" version="1.5.0" targetFramework="net452" />
   <package id="MahApps.Metro.IconPacks.Entypo" version="1.9.0" targetFramework="net452" />
   <package id="MahApps.Metro.IconPacks.Modern" version="1.9.0" targetFramework="net452" />
